refactor(#18): complete refactoring.
