+++ src/main/scala/jp/ed/nnn/movieplayer/MoviePlayer.scala
@@ -1,0 +1,60 @@
+package jp.ed.nnn.movieplayer
+
+import javafx.beans.value.{ChangeListener, ObservableValue}
+import javafx.scene.control.{Label, TableView}
+import javafx.scene.media.{MediaPlayer, MediaView}
+import javafx.util.Duration
+
+object MoviePlayer {
+
+  def play(movie: Movie, tableView: TableView[Movie], mediaView: MediaView, timeLabel: Label): Unit = {
+    if (mediaView.getMediaPlayer != null) {
+      val oldPlayer = mediaView.getMediaPlayer
+      oldPlayer.stop()
+      oldPlayer.dispose()
+    }
+
+    val mediaPlayer = new MediaPlayer(movie.media)
+    mediaPlayer.currentTimeProperty().addListener(new ChangeListener[Duration] {
+      override def changed(observable: ObservableValue[_ <: Duration], oldValue: Duration, newValue: Duration): Unit =
+        timeLabel.setText(formatTime(mediaPlayer.getCurrentTime, mediaPlayer.getTotalDuration))
+    })
+    mediaPlayer.setOnReady(new Runnable {
+      override def run(): Unit =
+        timeLabel.setText(formatTime(mediaPlayer.getCurrentTime, mediaPlayer.getTotalDuration))
+    })
+    mediaPlayer.setOnEndOfMedia(new Runnable {
+      override def run(): Unit = playNext(tableView, mediaView, timeLabel)
+    })
+
+    mediaView.setMediaPlayer(mediaPlayer)
+    mediaPlayer.setRate(1.25)
+    mediaPlayer.play()
+  }
+
+  sealed trait Track
+  object Pre extends Track
+  object Next extends Track
+
+  private[this] def playAt(track: Track, tableView: TableView[Movie], mediaView: MediaView, timeLabel: Label): Unit = {
+    val selectionModel = tableView.getSelectionModel
+    if (selectionModel.isEmpty) return
+    val index = selectionModel.getSelectedIndex
+    val changeIndex = track match {
+      case Pre => (tableView.getItems.size() + index - 1) % tableView.getItems.size()
+      case Next => (index + 1) % tableView.getItems.size()
+    }
+    selectionModel.select(changeIndex)
+    val movie = selectionModel.getSelectedItem
+    play(movie, tableView, mediaView, timeLabel)
+  }
+
+  def playPre(tableView: TableView[Movie], mediaView: MediaView, timeLabel: Label): Unit = {
+    playAt(Pre, tableView, mediaView, timeLabel)
+  }
+
+  def playNext(tableView: TableView[Movie], mediaView: MediaView, timeLabel: Label): Unit = {
+    playAt(Next, tableView, mediaView, timeLabel)
+  }
+
+}
