+++ src/main/scala/jp/ed/nnn/movieplayer/SizeConstants.scala
@@ -1,0 +1,10 @@
+package jp.ed.nnn.movieplayer
+
+object SizeConstants {
+
+  val mediaViewFitWidth = 800
+  val mediaViewFitHeight = 450
+  val toolBarMinHeight = 50
+  val tableMinWidth = 300
+
+}
