+++ src/main/scala/jp/ed/nnn/movieplayer/ToolbarCreator.scala
@@ -1,0 +1,104 @@
+package jp.ed.nnn.movieplayer
+
+import javafx.event.{ActionEvent, EventHandler}
+import javafx.geometry.Pos
+import javafx.scene.control.{Button, Label, TableView}
+import javafx.scene.image.{Image, ImageView}
+import javafx.scene.input.MouseEvent
+import javafx.scene.layout.HBox
+import javafx.scene.media.MediaView
+import javafx.stage.Stage
+import javafx.util.Duration
+
+import jp.ed.nnn.movieplayer.SizeConstants._
+
+object ToolbarCreator {
+
+  def create(mediaView: MediaView, tableView: TableView[Movie], timeLabel: Label, primaryStage: Stage): HBox = {
+    val toolBar = new HBox()
+    toolBar.setMinHeight(toolBarMinHeight)
+    toolBar.setAlignment(Pos.CENTER)
+    toolBar.setStyle("-fx-background-color: Black")
+
+    // first Button
+    val firstButton = createButton("images/first.png", new EventHandler[ActionEvent] {
+      override def handle(event: ActionEvent): Unit =
+        if (mediaView.getMediaPlayer != null) {
+          MoviePlayer.playPre(tableView, mediaView, timeLabel)
+        }
+    })
+
+    // back button
+    val backButton = createButton("images/back.png", new EventHandler[ActionEvent] {
+      override def handle(event: ActionEvent): Unit =
+        if (mediaView.getMediaPlayer != null) {
+          mediaView.getMediaPlayer.seek(
+            mediaView.getMediaPlayer.getCurrentTime.subtract(new Duration(10000))
+          )
+        }
+    })
+
+    // play button
+    val playButton = createButton("images/play.png", new EventHandler[ActionEvent] {
+      override def handle(event: ActionEvent): Unit = {
+        val selectionMode = tableView.getSelectionModel
+        if (mediaView.getMediaPlayer != null && !selectionMode.isEmpty) {
+          mediaView.getMediaPlayer.play()
+        }
+      }
+    })
+
+    // pause button
+    val pauseButton = createButton("images/pause.png", new EventHandler[ActionEvent] {
+      override def handle(event: ActionEvent): Unit = {
+        if (mediaView.getMediaPlayer != null) mediaView.getMediaPlayer.pause()
+      }
+    })
+
+    // forward button
+    val forwardButton = createButton("images/forward.png", new EventHandler[ActionEvent] {
+      override def handle(event: ActionEvent): Unit = {
+        if (mediaView.getMediaPlayer != null) {
+          mediaView.getMediaPlayer.seek(
+            mediaView.getMediaPlayer.getCurrentTime.add(new Duration(10000))
+          )
+        }
+      }
+    })
+
+    // last Button
+    val lastButton = createButton("images/last.png", new EventHandler[ActionEvent] {
+      override def handle(event: ActionEvent): Unit =
+        if (mediaView.getMediaPlayer != null) {
+          MoviePlayer.playNext(tableView, mediaView, timeLabel)
+        }
+    })
+
+    // fullscreen Button
+    val fullscreenButton = createButton("images/fullscreen.png", new EventHandler[ActionEvent] {
+      override def handle(event: ActionEvent): Unit =
+        primaryStage.setFullScreen(true)
+    })
+
+    toolBar.getChildren.addAll(firstButton, backButton, playButton, pauseButton, forwardButton, lastButton, fullscreenButton, timeLabel)
+  }
+
+  private[this] def createButton(imagePath: String, eventHandler: EventHandler[ActionEvent]): Button = {
+    val buttonImage = new Image(getClass.getResourceAsStream(imagePath))
+    val button = new Button()
+    button.setGraphic(new ImageView(buttonImage))
+    button.setStyle("-fx-background-color: Black")
+    button.setOnAction(eventHandler)
+    button.addEventHandler(MouseEvent.MOUSE_ENTERED, new EventHandler[MouseEvent] {
+      override def handle(event: MouseEvent): Unit = {
+        button.setStyle("-fx-body-color: Black")
+      }
+    })
+    button.addEventHandler(MouseEvent.MOUSE_EXITED, new EventHandler[MouseEvent] {
+      override def handle(event: MouseEvent): Unit = {
+        button.setStyle("-fx-background-color: Black")
+      }
+    })
+    button
+  }
+}
